# In-memory caching for node image access multitenancy