#### What type of PR is this?
/kind feature

#### What this PR does / why we need it:
This PR adds a write-through caching layer between the kubelet's image pulls manager and the on-disk image pull records. The cache implements a fallback to the disk on cache miss. 

#### Which issue(s) this PR fixes:
Related-to #https://github.com/kubernetes/enhancements/issues/2535

#### Special notes for your reviewer:
_NOTE_: This PR relies, and is rebased, on benchmark code from https://github.com/kubernetes/kubernetes/pull/131864.

The benchmarks below show the comparison to direct access to image pull records on disk, and when the feature is disabled. We could probably improve the performance further by having the in-memory cache be authoritative source with an async records refetch, though the current fallback approach was chosen as more safer as it handles incomplete delegate record listing during the cache init.

Raw benchmark results:
[directfs_memcache_comparison_28GiB.txt](https://github.com/user-attachments/files/20366428/directfs_memcache_comparison_28GiB.txt)
[directfs_memcache_benchmark_28GiB.txt](https://github.com/user-attachments/files/20366427/directfs_memcache_benchmark_28GiB.txt)


1. **Feature enabled/disabled results (ran on a system with 28GiB memory)** - columns are the state of the feature
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/kubelet/images
cpu: Intel(R) Xeon(R) CPU E5-2673 v4 @ 2.30GHz
                                                                                │  Disabled   │               DirectFS               │              CachedFS               │
                                                                                │   sec/op    │   sec/op     vs base                 │   sec/op     vs base                │
ImagePullManager_CompareEnsureSecretPulledImages/Records=100/CacheHitRate=100-8   283.8µ ± 3%   689.0µ ± 3%  +142.77% (p=0.000 n=10)   408.3µ ± 2%  +43.85% (p=0.000 n=10)

                                                                                │   Disabled   │               DirectFS               │               CachedFS               │
                                                                                │     B/op     │     B/op      vs base                │     B/op      vs base                │
ImagePullManager_CompareEnsureSecretPulledImages/Records=100/CacheHitRate=100-8   39.22Ki ± 1%   61.44Ki ± 1%  +56.63% (p=0.000 n=10)   58.18Ki ± 2%  +48.32% (p=0.000 n=10)

                                                                                │  Disabled  │               DirectFS               │              CachedFS              │
                                                                                │ allocs/op  │  allocs/op   vs base                 │ allocs/op   vs base                │
ImagePullManager_CompareEnsureSecretPulledImages/Records=100/CacheHitRate=100-8   45.00 ± 0%   108.00 ± 0%  +140.00% (p=0.000 n=10)   58.00 ± 0%  +28.89% (p=0.000 n=10)
```
We can see that caching improves the performance greatly on 100% cache hits compared to accessing the files directly though still being slower than if the feature is disabled. This behavior is expected.

2. **Comparing the direct FS access to in-memory caching based on cache hit rate and number of records** - columns describe the access method
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/kubelet/images
cpu: Intel(R) Xeon(R) CPU E5-2673 v4 @ 2.30GHz
                                                                           │   DirectFS   │              CachedFS               │
                                                                           │    sec/op    │   sec/op     vs base                │
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=10-8      1.099m ± 2%   1.125m ± 2%   +2.37% (p=0.011 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=20-8      1.050m ± 3%   1.039m ± 2%        ~ (p=0.165 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=50-8      898.5µ ± 3%   805.2µ ± 4%  -10.39% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=75-8      789.2µ ± 3%   616.9µ ± 4%  -21.83% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=100-8     696.1µ ± 2%   408.1µ ± 2%  -41.38% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=10-8      1.137m ± 1%   1.131m ± 3%   -0.56% (p=0.043 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=20-8      1.089m ± 2%   1.040m ± 2%   -4.50% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=50-8      936.7µ ± 2%   779.5µ ± 2%  -16.79% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=75-8      812.1µ ± 2%   586.9µ ± 4%  -27.73% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=100-8     704.2µ ± 3%   409.0µ ± 2%  -41.92% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=10-8     1.138m ± 1%   1.094m ± 2%   -3.81% (p=0.001 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=20-8     1.058m ± 4%   1.024m ± 3%   -3.17% (p=0.023 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=50-8     931.4µ ± 2%   788.8µ ± 1%  -15.32% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=75-8     820.5µ ± 3%   591.1µ ± 3%  -27.95% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=100-8    684.4µ ± 3%   415.0µ ± 2%  -39.37% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=10-8     1.232m ± 3%   1.163m ± 1%   -5.62% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=20-8     1.175m ± 1%   1.078m ± 2%   -8.21% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=50-8    1001.3µ ± 3%   845.5µ ± 2%  -15.56% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=75-8     847.3µ ± 3%   659.2µ ± 3%  -22.19% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=100-8    722.4µ ± 2%   457.5µ ± 3%  -36.67% (p=0.000 n=10)
geomean                                                                       925.1µ        755.0µ       -18.40%

                                                                           │   DirectFS   │              CachedFS               │
                                                                           │     B/op     │     B/op      vs base               │
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=10-8     68.38Ki ± 2%   68.16Ki ± 2%       ~ (p=0.436 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=20-8     67.65Ki ± 3%   66.54Ki ± 1%  -1.65% (p=0.002 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=50-8     65.57Ki ± 2%   63.60Ki ± 3%  -3.01% (p=0.002 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=75-8     63.65Ki ± 2%   60.53Ki ± 2%  -4.90% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=100-8    61.39Ki ± 2%   57.40Ki ± 1%  -6.50% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=10-8     68.99Ki ± 4%   68.58Ki ± 2%       ~ (p=0.481 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=20-8     67.07Ki ± 2%   67.01Ki ± 2%       ~ (p=0.393 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=50-8     65.80Ki ± 2%   63.50Ki ± 2%  -3.50% (p=0.003 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=75-8     63.29Ki ± 3%   60.06Ki ± 3%  -5.11% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=100-8    61.69Ki ± 2%   57.76Ki ± 2%  -6.37% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=10-8    68.28Ki ± 2%   68.21Ki ± 2%       ~ (p=0.644 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=20-8    68.04Ki ± 2%   66.76Ki ± 3%       ~ (p=0.247 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=50-8    65.75Ki ± 2%   63.69Ki ± 2%  -3.14% (p=0.001 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=75-8    64.90Ki ± 3%   60.47Ki ± 2%  -6.83% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=100-8   61.56Ki ± 2%   57.98Ki ± 2%  -5.82% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=10-8    68.77Ki ± 2%   68.62Ki ± 1%       ~ (p=0.481 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=20-8    69.55Ki ± 4%   66.90Ki ± 2%  -3.81% (p=0.005 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=50-8    65.23Ki ± 3%   63.90Ki ± 4%       ~ (p=0.165 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=75-8    63.90Ki ± 2%   60.82Ki ± 3%  -4.81% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=100-8   61.80Ki ± 2%   58.45Ki ± 3%  -5.43% (p=0.000 n=10)
geomean                                                                      65.51Ki        63.33Ki       -3.33%

                                                                           │  DirectFS   │              CachedFS              │
                                                                           │  allocs/op  │ allocs/op   vs base                │
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=10-8      150.0 ± 0%   146.0 ± 1%   -2.67% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=20-8      146.0 ± 1%   135.5 ± 1%   -7.19% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=50-8      131.0 ± 1%   106.0 ± 2%  -19.08% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=75-8     119.50 ± 0%   82.00 ± 1%  -31.38% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=10/CacheHitRate=100-8    108.00 ± 0%   58.00 ± 0%  -46.30% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=10-8      150.5 ± 1%   146.0 ± 1%   -2.99% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=20-8      145.0 ± 1%   136.0 ± 1%   -6.21% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=50-8      131.0 ± 1%   106.5 ± 1%  -18.70% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=75-8     119.00 ± 1%   81.00 ± 1%  -31.93% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=50/CacheHitRate=100-8    108.00 ± 1%   58.00 ± 0%  -46.30% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=10-8     150.0 ± 1%   146.0 ± 1%   -2.67% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=20-8     146.0 ± 1%   136.5 ± 2%   -6.51% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=50-8     131.0 ± 1%   107.0 ± 2%  -18.32% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=75-8    120.00 ± 1%   82.00 ± 2%  -31.67% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=100/CacheHitRate=100-8   108.00 ± 0%   58.00 ± 0%  -46.30% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=10-8     150.0 ± 1%   146.0 ± 1%   -2.67% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=20-8     145.0 ± 1%   137.0 ± 1%   -5.52% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=50-8     131.0 ± 1%   107.0 ± 1%  -18.32% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=75-8    119.50 ± 0%   82.00 ± 1%  -31.38% (p=0.000 n=10)
ImagePullManageWithEnsureSecretPulledImages/Records=500/CacheHitRate=100-8   108.00 ± 0%   58.00 ± 0%  -46.30% (p=0.000 n=10)
geomean                                                                       129.9        100.1       -22.91%
```
As expected, the overhead of the write-through cache shows at low number of records and low cache hit rates, but in general the performance improvement is apparent in scenarios that are more likely in the real world - above 50% hit rate and 50-70 records. The number of allocations also gets much lower, most likely because we don't need to encode/decode the resources too often.

#### Does this PR introduce a user-facing change?
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:
```docs
- [KEP]: https://github.com/kubernetes/enhancements/issues/2535
```

/cc liggitt enj 
/sig node
/sig auth