@stlaz 2025-06-04
/retitle [WIP] In-memory caching for node image access multitenancy
The current implementation loads all the file-based records in memory on init. @benjaminapetersen reminded me that this is perhaps a bit naive implementation and that we may want to limit the number of records in memory. I'll add some kind of an LRU mechanism to the cache.